refactor: Extract database models and SQL code to separate model.py module
- Created new backend/model.py module containing:
  * User model with password hashing methods
  * OAuthProvider model for OAuth configuration
  * OAuthAccount model for OAuth account linking
  * SQLAlchemy and JWT manager instances
  * Proper docstrings and __repr__ methods

- Refactored backend/app.py:
  * Removed model definitions and moved to model.py
  * Updated imports to use models from model.py
  * Changed db and jwt initialization to use init_app() pattern
  * Kept JWT handlers that depend on Redis in app.py
  * Maintained all existing functionality

- Benefits of refactoring:
  * Better separation of concerns
  * Improved code organization
  * Easier model maintenance and testing
  * Cleaner app.py focused on routes and business logic
  * Models can be imported independently

This refactoring improves code structure while maintaining
all existing functionality and API endpoints.
